statusfooter into separate component
